+++ packages/orq-rc/src/orq_ai_sdk/memorystores.py
@@ -289,18 +289,14 @@ class MemoryStores(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "4XX", "5XX"],
+            error_status_codes=["4XX", "5XX"],
             retry_config=retry_config,
         )
 
-        response_data: Any = None
         if utils.match_response(http_res, "201", "application/json"):
             return unmarshal_json_response(
                 Optional[models.CreateMemoryStoreResponseBody], http_res
             )
-        if utils.match_response(http_res, "400", "application/json"):
-            response_data = unmarshal_json_response(models.HonoAPIErrorData, http_res)
-            raise models.HonoAPIError(response_data, http_res)
         if utils.match_response(http_res, "4XX", "*"):
             http_res_text = utils.stream_to_text(http_res)
             raise models.APIError("API error occurred", http_res, http_res_text)
@@ -393,18 +389,14 @@ class MemoryStores(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "4XX", "5XX"],
+            error_status_codes=["4XX", "5XX"],
             retry_config=retry_config,
         )
 
-        response_data: Any = None
         if utils.match_response(http_res, "201", "application/json"):
             return unmarshal_json_response(
                 Optional[models.CreateMemoryStoreResponseBody], http_res
             )
-        if utils.match_response(http_res, "400", "application/json"):
-            response_data = unmarshal_json_response(models.HonoAPIErrorData, http_res)
-            raise models.HonoAPIError(response_data, http_res)
         if utils.match_response(http_res, "4XX", "*"):
             http_res_text = await utils.stream_to_text_async(http_res)
             raise models.APIError("API error occurred", http_res, http_res_text)