+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -63,5 +63,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-11-20T07:47:08.022Z")
+    created: Optional[datetime] = parse_datetime("2025-11-20T08:48:01.918Z")
     r"""The date and time the resource was created"""