+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -130,7 +130,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAG3PQC6FC4KPN6J7X3ZYQSZ"
+        "01KAG769PZMTEJRRMK95K9105H"
     )
     r"""The unique identifier of the data source"""