+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAG3PQCC00VJ42R0G2ZF2SEF"
+        "01KAG769Q298YV9VCN6NATDG1H"
     )
     r"""The unique identifier of the data source"""