+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAG3PQCY01GZSPW87ADQX0WS"
+        "01KAG769Q76RQMT9MVANEZE9H2"
     )
     r"""The unique identifier of the data source"""